settings: use new api for start menu
