+++ BlasII.Randomizer/BlasII.Randomizer.csproj
@@ -15,9 +15,9 @@
     <RepositoryUrl>https://github.com/BrandenEK/BlasII.Randomizer</RepositoryUrl>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
   </PropertyGroup>
-
+  
   <ItemGroup>
-    <PackageReference Include="BlasII.ModdingAPI" Version="0.7.0" />
+    <PackageReference Include="BlasII.ModdingAPI" Version="0.8.0" />
     <PackageReference Include="BlasII.ModdingReferences" Version="1.1.0.2" />
     <PackageReference Include="LogicParser" Version="1.1.0" />
   </ItemGroup>
