+++ BlasII.Randomizer/Properties/AssemblyInfo.cs
@@ -2,4 +2,4 @@ using BlasII.Randomizer;
 using MelonLoader;
 
 [assembly: MelonInfo(typeof(BlasII.Randomizer.Main), ModInfo.MOD_NAME, ModInfo.MOD_VERSION, ModInfo.MOD_AUTHOR)]
-[assembly: MelonOptionalDependencies("Modding API", "0.7.0", "Quality of Life", "1.2.0")]
+[assembly: MelonOptionalDependencies("Modding API", "0.8.0", "Quality of Life", "1.2.0")]
